A2Q2 q2_parser_model.py for debug works!

+++ a2/q2_initialization.py
@@ -25,8 +25,8 @@ def xavier_weight_init():
             out: tf.Tensor of specified shape sampled from the Xavier distribution.
         """
         ### YOUR CODE HERE
-        epsilon = tf.sqrt(np.product(shape) / np.sum(shape))
-        out = tf.random_uniform(shape = shape, dtype = tf.float64, minval = -epsilon, maxval = epsilon)
+        epsilon = np.sqrt(np.product(shape) / np.sum(shape))
+        out = tf.random_uniform(shape = shape, dtype = tf.float32, minval = -epsilon, maxval = epsilon)
         ### END YOUR CODE
         return out
     # Returns defined initializer function.
